# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenbookLM Core API",
    "description": "The public core of OpenbookLM: notebooks, sources, chunks, notes, memories, chat and retrieval metrics. Event-stream semantics are specified separately in docs/contracts/sse-protocol-v1.md.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/notebooks/{id}/chat": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "GET /api/notebooks/{id}/chat - Get paginated chat history.",
        "operationId": "get_chat_history_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of chat history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "POST /api/notebooks/{id}/chat - Send a message and get streaming response.",
        "operationId": "send_message_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server-sent event stream. Ordering, terminal events and cancellation are specified in docs/contracts/sse-protocol-v1.md, which OpenAPI does not model.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ChatEvent"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Denied by the entitlement policy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "DELETE /api/notebooks/{id}/chat - Clear chat history.",
        "operationId": "clear_chat_history_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion acknowledgement",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/teaching-modes": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "GET /api/teaching-modes - List available teaching modes.",
        "operationId": "list_teaching_modes",
        "responses": {
          "200": {
            "description": "Available teaching modes and the default",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingModesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatChunk": {
        "type": "object",
        "description": "A partial text delta from the provider.",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
      "ChatCitationRef": {
        "type": "object",
        "description": "An inline `[N]` citation marker, emitted as soon as it appears in the text.",
        "required": [
          "index",
          "source_id"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "source_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ChatEvent": {
        "oneOf": [
          {
            "type": "object",
            "description": "Text delta. Emitted zero or more times.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatChunk",
                "description": "Text delta. Emitted zero or more times."
              },
              "event": {
                "type": "string",
                "enum": [
                  "chunk"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Pipeline stage indicator.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatThinking",
                "description": "Pipeline stage indicator."
              },
              "event": {
                "type": "string",
                "enum": [
                  "thinking"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Prompt assembly notice.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatSystem",
                "description": "Prompt assembly notice."
              },
              "event": {
                "type": "string",
                "enum": [
                  "system"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Non-fatal retrieval warning.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatWarning",
                "description": "Non-fatal retrieval warning."
              },
              "event": {
                "type": "string",
                "enum": [
                  "warning"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Inline citation marker.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatCitationRef",
                "description": "Inline citation marker."
              },
              "event": {
                "type": "string",
                "enum": [
                  "citation"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Resolved citation set.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatCitations",
                "description": "Resolved citation set."
              },
              "event": {
                "type": "string",
                "enum": [
                  "citations"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Retrieval quality metrics.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatMetrics",
                "description": "Retrieval quality metrics."
              },
              "event": {
                "type": "string",
                "enum": [
                  "metrics"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Follow-up questions. Always emitted before [`ChatEvent::Done`].",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatFollowUpSuggestions",
                "description": "Follow-up questions. Always emitted before [`ChatEvent::Done`]."
              },
              "event": {
                "type": "string",
                "enum": [
                  "follow_up_suggestions"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Terminal success.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatDone",
                "description": "Terminal success."
              },
              "event": {
                "type": "string",
                "enum": [
                  "done"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Terminal failure.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatError",
                "description": "Terminal failure."
              },
              "event": {
                "type": "string",
                "enum": [
                  "error"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Terminal server shutdown.",
            "required": [
              "data",
              "event"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/ChatShutdown",
                "description": "Terminal server shutdown."
              },
              "event": {
                "type": "string",
                "enum": [
                  "shutdown"
                ]
              }
            }
          }
        ],
        "description": "Every event the chat stream can emit, in `v1` of the protocol.\n\nClients must ignore unknown variants without terminating the stream:\nadditive variants are not a breaking change."
      },
      "ChatFollowUpSuggestions": {
        "type": "object",
        "description": "Suggested follow-up questions.",
        "required": [
          "suggestions"
        ],
        "properties": {
          "suggestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ChatHistoryResponse": {
        "type": "object",
        "description": "Response for paginated chat history.",
        "required": [
          "messages",
          "total",
          "offset",
          "limit",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            }
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ChatMessageResponse": {
        "type": "object",
        "description": "Chat message response DTO.\n\nLives in the service layer (not API) because it's shared by both\nthe API handlers and the export service.",
        "required": [
          "id",
          "notebook_id",
          "role",
          "content",
          "citations",
          "created_at"
        ],
        "properties": {
          "citations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Citation"
            }
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "feedback": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "notebook_id": {
            "type": "string",
            "format": "uuid"
          },
          "rag_log_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "role": {
            "type": "string"
          },
          "session_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
...
      "ChatThinking": {
        "type": "object",
        "description": "Thinking indicator payload.",
        "required": [
          "stage"
        ],
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/ThinkingStage"
          }
        }
      },
      "ChatWarning": {
        "type": "object",
        "description": "Warning payload.",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/WarningKind"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "retry_after": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "description": "Request for sending a chat message.",
        "required": [
          "message"
        ],
        "properties": {
          "locale": {
            "type": [
              "string",
              "null"
            ],
            "description": "UI locale for generating suggestions in the correct language (e.g. \"fr\", \"en\")"
          },
          "max_context_chunks": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "model": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional model override"
          },
          "provider": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional LLM provider (e.g., \"mistral\", \"anthropic\")"
          },
          "teaching_mode": {
            "$ref": "#/components/schemas/TeachingMode",
            "description": "Teaching mode for pedagogical adaptation"
          }
        }
      },
...
      "TeachingMode": {
        "type": "string",
        "description": "Teaching mode for pedagogical adaptation.",
        "enum": [
          "flash",
          "deep",
          "quiz",
          "glossary",
          "summary",
          "timeline"
        ]
      },
...
      "TeachingModesResponse": {
        "type": "object",
        "description": "Response for teaching modes list.",
        "required": [
          "modes",
          "default"
        ],
        "properties": {
          "default": {
            "type": "string"
          },
          "modes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TeachingModeInfo"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token verified by the composed identity adapter."
      }
    }
  },
...
}